requirements.txt
streamlit>=1.24.0
numpy>=1.20.0
pandas>=1.3.0
matplotlib>=3.5.0
seaborn>=0.11.0
scipy>=1.7.0
scikit-learn>=1.0.0
numba>=0.56.0
tensorflow>=2.8.0
joblib>=1.1.0
plotly>=5.5.0
yfinance>=0.1.70
requests>=2.28.0
black>=23.1.0
pytest>=7.0.0
jupyter>=1.0.0
ipython>=8.0.0
pylint>=2.17.0
mypy>=1.3.0
influxdb-client>=1.36.0
questdb>=1.1.0
ib_insync>=0.9.85
alpaca-py>=0.8.0
redis>=4.5.0
celery>=5.3.0
prometheus_client>=0.17.0
pyyaml>=6.0
python-dotenv>=1.0.0
fastapi>=0.95.0
uvicorn>=0.22.0
docker>=6.1.0 